backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
streamlit==1.28.1
pytest==7.4.3
requests==2.31.0
email-validator==2.1.0
aiofiles==23.2.1
python-magic==0.4.27  # Pour la détection de type MIME
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
streamlit==1.28.1
pytest==7.4.3
requests==2.31.0
email-validator==2.1.0
aiofiles==23.2.1
python-magic==0.4.27  # Pour la détection de type MIME
